added public key controller
