+++ request-generator/src/db.json
@@ -1,12 +1,5 @@
 {
   "public_keys": [
-    {
-      "pem": {
-        "kty": "RSA",
-        "n": "uDJMn6pBJm2xqvmRHU0QuZg-ug6CXz5tEf_h0GI1e902xv-_WcptoJ_5v--bN3hOGxP1juVCK8KQ-8hWzHeIOY8T9o3QWC7YN6uWyMFtBFSz1IBRkzDfNL8w-roBf1eEFsbn-jrKAptxfZcQu9009fa2gwXMN2ZF18cA5i0HXa73piKZhjxKyS8SeS7uhAnVJGSjs1OVBTW_R3sJh6rWKL0eX5_H_pykkBXUnEMRyUJEO7isn0qE-_hQcU0AbdLBycq5AUngN590iWwHd7UqE3ENRQNHujgWBR_IuO6POrWib88V6rcjIZztwJ3hVfOFCb13D17AZW6sXSh4fvhIhQ",
-        "e": "AQAB"
-      },
-      "id": "SA3DcGdl45hr06S-VDTOkx4bDlKKodlu-pyLdj0W8qI"
-    }
+
   ]
 }
